Fix validation and compatibility regressions
