+++ src/agilab/apps-pages/view_live_artifacts/src/view_live_artifacts/view_live_artifacts.py
@@ -366,6 +366,18 @@ def _render_controls(env: AgiEnv, active_app_path: Path) -> tuple[Path, tuple[st
     return selected_root, parse_patterns(pattern_value), max_files, live_refresh, interval_seconds
 
 
+def _render_code_block(body: str, *, language: str, height: int | None = None) -> None:
+    if height is None:
+        st.code(body, language=language)
+        return
+    try:
+        st.code(body, language=language, height=height)
+    except TypeError as exc:
+        if "height" not in str(exc):
+            raise
+        st.code(body, language=language)
+
+
 def _render_preview(records: tuple[ArtifactRecord, ...]) -> None:
     options = [record.relative_path for record in records]
     if not options:
@@ -376,13 +388,13 @@ def _render_preview(records: tuple[ArtifactRecord, ...]) -> None:
     st.caption(f"{selected.relative_path} · {format_bytes(selected.size)} · {selected.mtime_iso}")
     if preview.error:
         st.error("Preview unavailable.")
-        st.code(preview.error, language="text", height=400)
+        _render_code_block(preview.error, language="text", height=400)
     elif preview.kind == "json":
         st.json(preview.value, expanded=False)
     elif preview.kind == "image":
         st.image(str(preview.value), caption=selected.relative_path)
     elif preview.kind == "text":
-        st.code(str(preview.value), language="text", height=400)
+        _render_code_block(str(preview.value), language="text", height=400)
         if preview.truncated:
             st.caption("Showing the latest portion of this file.")
     else:
