+++ src/agilab/lib/agi-app-pytorch-playground/src/agi_app_pytorch_playground/project/pytorch_playground_project/src/pytorch_playground/__init__.py
@@ -13,7 +13,16 @@ from .runtime.app_args import (
     merge_args,
     to_playground_config,
 )
-from .runtime.reduction import PYTORCH_PLAYGROUND_REDUCE_CONTRACT
+from .reduction import (
+    PYTORCH_PLAYGROUND_REDUCE_CONTRACT,
+    REDUCE_ARTIFACT_FILENAME_TEMPLATE,
+    REDUCE_ARTIFACT_NAME,
+    REDUCER_NAME,
+    build_reduce_artifact,
+    partial_from_summary,
+    reduce_artifact_path,
+    write_reduce_artifact,
+)
 
 _MANAGER_EXPORTS = {"PytorchPlayground", "PytorchPlaygroundApp"}
 
@@ -39,9 +48,16 @@ __all__ = [
     "PYTORCH_PLAYGROUND_REDUCE_CONTRACT",
     "PytorchPlaygroundArgs",
     "PytorchPlaygroundArgsTD",
+    "REDUCE_ARTIFACT_FILENAME_TEMPLATE",
+    "REDUCE_ARTIFACT_NAME",
+    "REDUCER_NAME",
+    "build_reduce_artifact",
     "dump_args",
     "ensure_defaults",
     "load_args",
     "merge_args",
+    "partial_from_summary",
+    "reduce_artifact_path",
     "to_playground_config",
+    "write_reduce_artifact",
 ]
