+++ src/agilab/lib/agi-gui/src/agi_gui/compat/module_alias.py
@@ -19,5 +19,6 @@ def alias_agi_env_module(target_name: str, source_name: str) -> ModuleType:
     """Expose an existing agi_env UI module under the agi_gui namespace."""
 
     module = import_module(source_name)
+    sys.modules[source_name] = module
     sys.modules[target_name] = module
     return module
