+++ src/agilab/reuse/__init__.py
@@ -1,0 +1,5 @@
+"""Reuse-analysis helpers for AGILAB."""
+
+from __future__ import annotations
+
+__all__: list[str] = []
