+++ test/test_tools_surface_contract.py
@@ -6,7 +6,7 @@ from pathlib import Path
 # Current public release tooling keeps repository, release, HF sync, shared
 # go/no-go gates, and guardrail entrypoints top-level so workflows can call them
 # directly.
-TOOLS_SURFACE_BUDGET = 164
+TOOLS_SURFACE_BUDGET = 187
 
 
 def test_top_level_tools_surface_stays_within_budget() -> None:
